move test helpers from chain module to section module
